test(cron): split session-target default tests to satisfy max-lines
