+++ src/cron/normalize.test.ts
@@ -433,34 +433,6 @@ describe("normalizeCronJobCreate", () => {
     expect(delivery.mode).toBe("announce");
   });
 
-  it("defaults agentTurn jobs with session context to current announce jobs", () => {
-    const normalized = normalizeCronJobCreate(
-      {
-        name: "agent turn current default",
-        schedule: { kind: "every", everyMs: 60_000 },
-        payload: { kind: "agentTurn", message: "hello" },
-      },
-      { sessionContext: { sessionKey: "agent:main:discord:channel:ops" } },
-    );
-    expect(normalized).toMatchObject({
-      sessionTarget: "current",
-      sessionKey: "agent:main:discord:channel:ops",
-      delivery: { mode: "announce" },
-    });
-  });
-
-  it("downgrades the agentTurn current default to isolated without session context", () => {
-    const normalized = normalizeCronJobCreate({
-      name: "agent turn isolated fallback",
-      schedule: { kind: "every", everyMs: 60_000 },
-      payload: { kind: "agentTurn", message: "hello" },
-    });
-    expect(normalized).toMatchObject({
-      sessionTarget: "isolated",
-      delivery: { mode: "announce" },
-    });
-  });
-
   it("defaults command payloads to isolated announce jobs", () => {
     const normalized = normalizeCronJobCreate({
       name: "command default",
